azrepos: add non-PAT based auth mode to AzRepos provider
Add a new setting to the Azure Repos provider that instructs GCM to
return the Azure access token directly, rather than use that token to
generate a new Azure DevOps Personal Access Token.

At the moment the only indication as to what account a user wants to use
is via the userinfo part of the remote URL for visualstudio.com style
URLs. This must be manually set.

The default configuration is still to continue to use PATs, for now.

+++ src/shared/Microsoft.AzureRepos/AzureReposHostProvider.cs
@@ -73,43 +73,64 @@ namespace Microsoft.AzureRepos
 
         public async Task<ICredential> GetCredentialAsync(InputArguments input)
         {
-            Uri remoteUri = input.GetRemoteUri();
-            string service = GetServiceName(remoteUri);
-            string account = GetAccountNameForCredentialQuery(input);
-
-            _context.Trace.WriteLine($"Looking for existing credential in store with service={service} account={account}...");
-
-            ICredential credential = _context.CredentialStore.Get(service, account);
-            if (credential == null)
+            // We should not allow unencrypted communication and should inform the user
+            if (StringComparer.OrdinalIgnoreCase.Equals(input.Protocol, "http"))
             {
-                _context.Trace.WriteLine("No existing credentials found.");
-
-                // No existing credential was found, create a new one
-                _context.Trace.WriteLine("Creating new credential...");
-                credential = await GenerateCredentialAsync(input);
-                _context.Trace.WriteLine("Credential created.");
-            }
-            else
-            {
-                _context.Trace.WriteLine("Existing credential found.");
+                throw new Exception("Unencrypted HTTP is not supported for Azure Repos. Ensure the repository remote URL is using HTTPS.");
             }
 
-            return credential;
+            // Include any user from the URL/request here so that we may use it as an override
+            // for user account lookups when getting Azure Access Tokens.
+            Uri remoteUri = input.GetRemoteUri(includeUser: true);
+
+            if (IsPersonalAccessTokenMode())
+            {
+                string service = GetServiceName(input);
+                string account = GetAccountNameForCredentialQuery(input);
+
+                _context.Trace.WriteLine($"Looking for existing PAT in store with service={service} account={account}...");
+
+                ICredential credential = _context.CredentialStore.Get(service, account);
+                if (credential == null)
+                {
+                    _context.Trace.WriteLine("No existing PAT found.");
+
+                    // No existing credential was found, create a new one
+                    _context.Trace.WriteLine("Creating new PAT...");
+                    credential = await GeneratePersonalAccessTokenAsync(remoteUri);
+                    _context.Trace.WriteLine("PAT created.");
+                }
+                else
+                {
+                    _context.Trace.WriteLine("Existing PAT found.");
+                }
+
+                return credential;
+            }
+
+            var azureResult = await GetAzureAccessTokenAsync(remoteUri);
+            return new GitCredential(azureResult.AccountUpn, azureResult.AccessToken);
         }
 
         public Task StoreCredentialAsync(InputArguments input)
         {
-            Uri remoteUri = input.GetRemoteUri();
-            string service = GetServiceName(remoteUri);
+            if (IsPersonalAccessTokenMode())
+            {
+                string service = GetServiceName(input);
 
-            // We always store credentials against the given username argument for
-            // both vs.com and dev.azure.com-style URLs.
-            string account = input.UserName;
+                // We always store credentials against the given username argument for
+                // both vs.com and dev.azure.com-style URLs.
+                string account = input.UserName;
 
-            // Add or update the credential in the store.
-            _context.Trace.WriteLine($"Storing credential with service={service} account={account}...");
-            _context.CredentialStore.AddOrUpdate(service, account, input.Password);
-            _context.Trace.WriteLine("Credential was successfully stored.");
+                // Add or update the credential in the store.
+                _context.Trace.WriteLine($"Storing credential with service={service} account={account}...");
+                _context.CredentialStore.AddOrUpdate(service, account, input.Password);
+                _context.Trace.WriteLine("Credential was successfully stored.");
+            }
+            else
+            {
+                _context.Trace.WriteLine("Nothing to store when PAT mode is disabled.");
+            }
 
             return Task.CompletedTask;
         }
@@ -117,22 +138,30 @@ namespace Microsoft.AzureRepos
         public Task EraseCredentialAsync(InputArguments input)
         {
             Uri remoteUri = input.GetRemoteUri();
-            string service = GetServiceName(remoteUri);
-            string account = GetAccountNameForCredentialQuery(input);
+            string orgName = UriHelpers.GetOrganizationName(remoteUri);
 
-            // Try to locate an existing credential
-            _context.Trace.WriteLine($"Erasing stored credential in store with service={service} account={account}...");
-            if (_context.CredentialStore.Remove(service, account))
+            if (IsPersonalAccessTokenMode())
             {
-                _context.Trace.WriteLine("Credential was successfully erased.");
+                string service = GetServiceName(input);
+                string account = GetAccountNameForCredentialQuery(input);
+
+                // Try to locate an existing credential
+                _context.Trace.WriteLine($"Erasing stored credential in store with service={service} account={account}...");
+                if (_context.CredentialStore.Remove(service, account))
+                {
+                    _context.Trace.WriteLine("Credential was successfully erased.");
+                }
+                else
+                {
+                    _context.Trace.WriteLine("No credential was erased.");
+                }
             }
             else
             {
-                _context.Trace.WriteLine("No credential was erased.");
+                _context.Trace.WriteLine("Nothing to erase when PAT mode is disabled.");
             }
 
             // Clear the authority cache in case this was the reason for failure
-            string orgName = UriHelpers.GetOrganizationName(remoteUri);
             _authorityCache.EraseAuthority(orgName);
 
             return Task.CompletedTask;
@@ -144,17 +173,32 @@ namespace Microsoft.AzureRepos
             base.ReleaseManagedResources();
         }
 
-        private async Task<ICredential> GenerateCredentialAsync(InputArguments input)
+        private async Task<ICredential> GeneratePersonalAccessTokenAsync(Uri remoteUri)
         {
             ThrowIfDisposed();
 
-            // We should not allow unencrypted communication and should inform the user
-            if (StringComparer.OrdinalIgnoreCase.Equals(input.Protocol, "http"))
-            {
-                throw new Exception("Unencrypted HTTP is not supported for Azure Repos. Ensure the repository remote URL is using HTTPS.");
-            }
+            IMicrosoftAuthenticationResult azureResult = await GetAzureAccessTokenAsync(remoteUri);
 
-            Uri remoteUri = input.GetRemoteUri();
+            Uri orgUri = UriHelpers.CreateOrganizationUri(remoteUri, out _);
+
+            // Ask the Azure DevOps instance to create a new PAT
+            var patScopes = new[]
+            {
+                AzureDevOpsConstants.PersonalAccessTokenScopes.ReposWrite,
+                AzureDevOpsConstants.PersonalAccessTokenScopes.ArtifactsRead
+            };
+            _context.Trace.WriteLine($"Creating Azure DevOps PAT with scopes '{string.Join(", ", patScopes)}'...");
+            string pat = await _azDevOps.CreatePersonalAccessTokenAsync(
+                orgUri,
+                azureResult.AccessToken,
+                patScopes);
+            _context.Trace.WriteLineSecrets("PAT created. PAT='{0}'", new object[] {pat});
+
+            return new GitCredential(azureResult.AccountUpn, pat);
+        }
+
+        private async Task<IMicrosoftAuthenticationResult> GetAzureAccessTokenAsync(Uri remoteUri)
+        {
             Uri orgUri = UriHelpers.CreateOrganizationUri(remoteUri, out string orgName);
 
             _context.Trace.WriteLine($"Determining Microsoft Authentication authority for Azure DevOps organization '{orgName}'...");
@@ -168,6 +212,23 @@ namespace Microsoft.AzureRepos
             }
             _context.Trace.WriteLine($"Authority is '{authAuthority}'.");
 
+            //
+            // If the remote URI is a classic "*.visualstudio.com" remote and we have a user specified in the URI
+            // then take that as the current AAD/MSA user.
+            //
+            // For "dev.azure.com" style remote we cannot use the user info part of the URI because this has been
+            // hacked to be the Azure DevOps organization name instead (not an actual username).
+            //
+            string userName = null;
+            if (UriHelpers.IsVisualStudioComHost(remoteUri.Host) &&
+                remoteUri.TryGetUserInfo(out userName, out _) &&
+                !string.IsNullOrWhiteSpace(userName))
+            {
+                _context.Trace.WriteLine($"Using username as specified in remote URL '{remoteUri}'.");
+            }
+
+            _context.Trace.WriteLine(string.IsNullOrWhiteSpace(userName) ? "No user found." : $"User is '{userName}'.");
+
             // Get an AAD access token for the Azure DevOps SPS
             _context.Trace.WriteLine("Getting Azure AD access token...");
             IMicrosoftAuthenticationResult result = await _msAuth.GetTokenAsync(
@@ -175,24 +236,11 @@ namespace Microsoft.AzureRepos
                 GetClientId(),
                 GetRedirectUri(),
                 AzureDevOpsConstants.AzureDevOpsDefaultScopes,
-                null);
+                userName);
             _context.Trace.WriteLineSecrets(
                 $"Acquired Azure access token. Account='{result.AccountUpn}' Token='{{0}}'", new object[] {result.AccessToken});
 
-            // Ask the Azure DevOps instance to create a new PAT
-            var patScopes = new[]
-            {
-                AzureDevOpsConstants.PersonalAccessTokenScopes.ReposWrite,
-                AzureDevOpsConstants.PersonalAccessTokenScopes.ArtifactsRead
-            };
-            _context.Trace.WriteLine($"Creating Azure DevOps PAT with scopes '{string.Join(", ", patScopes)}'...");
-            string pat = await _azDevOps.CreatePersonalAccessTokenAsync(
-                orgUri,
-                result.AccessToken,
-                patScopes);
-            _context.Trace.WriteLineSecrets("PAT created. PAT='{0}'", new object[] {pat});
-
-            return new GitCredential(result.AccountUpn, pat);
+            return result;
         }
 
         private string GetClientId()
@@ -243,9 +291,17 @@ namespace Microsoft.AzureRepos
         /// Users that need to clone a repository from Azure Repos against the full path therefore must
         /// use the vs.com-style remote URL and not the dev.azure.com one.
         /// </remarks>
-        private static string GetServiceName(Uri remoteUri)
+        private static string GetServiceName(InputArguments input)
         {
-            if (UriHelpers.IsDevAzureComHost(remoteUri.Host))
+            if (!input.TryGetHostAndPort(out string hostName, out _))
+            {
+                throw new InvalidOperationException("Failed to parse host name and/or port");
+            }
+
+            Uri remoteUri = input.GetRemoteUri();
+
+            // dev.azure.com
+            if (UriHelpers.IsDevAzureComHost(hostName))
             {
                 // We can never store the new dev.azure.com-style URLs against the full path because
                 // we have forced the useHttpPath option to true to in order to retrieve the AzDevOps
@@ -254,7 +310,7 @@ namespace Microsoft.AzureRepos
             }
 
             // *.visualstudio.com
-            if (UriHelpers.IsVisualStudioComHost(remoteUri.Host))
+            if (UriHelpers.IsVisualStudioComHost(hostName))
             {
                 // If we're given the full path for an older *.visualstudio.com-style URL then we should
                 // respect that in the service name.
@@ -292,6 +348,27 @@ namespace Microsoft.AzureRepos
             }
 
             throw new InvalidOperationException("Host is not Azure DevOps.");
+        }
+
+        /// <summary>
+        /// Check if Azure DevOps Personal Access Tokens should be used or not.
+        /// </summary>
+        /// <returns>True if Personal Access Tokens should be used, false otherwise.</returns>
+        private bool IsPersonalAccessTokenMode()
+        {
+            // Keep PAT mode on by default whilst AT mode is being tested
+            const bool defaultValue = true;
+
+            if (_context.Settings.TryGetSetting(
+                AzureDevOpsConstants.EnvironmentVariables.PatMode,
+                KnownGitCfg.Credential.SectionName,
+                AzureDevOpsConstants.GitConfiguration.Credential.PatMode,
+                out string valueStr))
+            {
+                return valueStr.ToBooleanyOrDefault(defaultValue);
+            }
+
+            return defaultValue;
         }
 
         #endregion
